refactor: rename TCAIssueReporter to ComposableIssueReporter
BREAKING CHANGE: TCAIssueReporter has been renamed to ComposableIssueReporter
BREAKING CHANGE: configureTCAReporting() has been renamed to configureComposableReporting()

- Rename TCAIssueReporter.swift to ComposableIssueReporter.swift
- Update class name from TCAIssueReporter to ComposableIssueReporter
- Update method name from configureTCAReporting to configureComposableReporting
- Update documentation to reflect The Composable Architecture naming
- This change improves naming consistency with the LockmanComposable module
